Extract TraceEvent and DefaultFormatter to separate files
- Create TraceEvent.kt with indent limit feature (40 chars max)
- Create DefaultFormatter.kt for formatting utilities
- Simplify CodePathTracerRule.kt by removing duplicate code
- Improve code organization and maintainability
