+++ code-path-tracer/src/main/kotlin/io/github/takahirom/codepathtracer/DefaultFormatter.kt
@@ -1,0 +1,11 @@
+package io.github.takahirom.codepathtracer
+
+/**
+ * Default formatter for trace events with configurable max length.
+ */
+object DefaultFormatter {
+    var defaultMaxLength = 30
+    
+    fun format(event: TraceEvent): String = event.defaultFormat(defaultMaxLength)
+    fun format(event: TraceEvent, maxLength: Int): String = event.defaultFormat(maxLength)
+}
